requirements.txt
boto3>=1.34.0
python-dotenv>=1.0.0
oci>=2.120.0
requests>=2.31.0
jinja2>=3.1.0
pyyaml>=6.0
python-dateutil>=2.8.0
msal>=1.28.0
flask>=3.0.0
bcrypt>=4.1.0
